chore(cli): moved the commands to own files
